fix(MOR-2031): move shared TX resolver out of semantic/, harden acceptance test
Review findings on PR #2845 (Agent Review: BLOCKED at ef6a2045):

- B5: resolveTxFeedbackState/stringField were a production VALUE import
  from presentation/ into semantic/, reversing the v3 ADR's one-way
  dependency direction (type-only is permitted, as established by
  projection.ts; a runtime value import is not). Relocated both, plus
  their types (TxFeedbackRail/TxKeyTreatment/TxFeedbackTone/
  TxFeedbackState), into presentation/languages/tx-feedback-state.ts. The
  one remaining semantic/ reference is a type-only import of
  TxSessionState.
- B1: the "third design language" acceptance test only checked output
  equality, so a fully hand-duplicated private copy of the decision passed
  all 22 assertions unchanged. Added spy-based assertions that fieldline's,
  studioline's, and the third language's renderers each actually call the
  shared resolveTxFeedbackState. Verified by hand that the assertion goes
  red against a hand-duplicated copy and stays green against the real,
  legitimate renderer.
- B2/B3/B4: corrected four false prose statements — a stale
  "character-for-character identical" claim re-introduced in the test
  header, an unqualified "every fixture" totality claim, and a resolver
  comment claiming Object.prototype-member session names fall through to
  the doubt branch when they are actually caught earlier.
- Non-blocking word-fixes made while touching the same lines: a miscounted
  "a third time" enumeration claim, and two adjacent comments that
  contradicted each other describing the same branch. The finiteNumber
  duplication across four renderer files is reported in the PR body for a
  follow-up ticket rather than fixed here.
- Declared the prototype-chain behaviour change in the PR body as the fix
  it is: base walked the prototype chain and produced malformed
  descriptors for five specific session inputs; head fails closed to
  doubt. Unreachable in production, but previously undeclared.

No descriptor-behaviour change beyond the already-present, now-declared
prototype-chain fix. Moved code is otherwise unchanged except two comments
and one internal helper's reused lookup table (SESSION_LABEL ->
KEY_TREATMENT; both have the same five own keys, so behaviour is
identical).
