+++ frontend/src/presentation/languages/studioline/state-feedback-renderer.ts
@@ -9,12 +9,12 @@
  * `txSessionState()`, which are themselves derived from the App-owned
  * `TxAuthoritySnapshot` — plus `fault` and `keyBlocked`. It never reads a raw
  * `ptt`. The fail-closed DECISION over those four fields is
- * `semantic/rx-tx-surface`'s `resolveTxFeedbackState` (MOR-2031): it takes a
- * named struct of exactly those four primitives, never a whole
- * `RendererViewModel`, so a fifth field cannot be smuggled in without
- * changing that signature. Unrecognised values fail CLOSED to the doubt
- * rail, never to the quiet RX one: "nothing is happening" is the dangerous
- * claim.
+ * `resolveTxFeedbackState` (MOR-2031, `presentation/languages/tx-feedback-
+ * state.ts`): it takes a named struct of exactly those four primitives,
+ * never a whole `RendererViewModel`, so a fifth field cannot be smuggled in
+ * without changing that signature. Unrecognised values fail CLOSED to the
+ * doubt rail, never to the quiet RX one: "nothing is happening" is the
+ * dangerous claim.
  *
  * Colour is never the sole channel — every state carries a distinct rail
  * THICKNESS and a distinct text label, so the state survives forced-colors
@@ -29,7 +29,7 @@
 import type { DesignLanguageTokens, RendererViewModel } from '../contract';
 import {
   resolveTxFeedbackState, stringField, type TxFeedbackRail, type TxKeyTreatment,
-} from '../../../semantic/rx-tx-surface';
+} from '../tx-feedback-state';
 import { STUDIOLINE_PALETTE } from './tokens';
 
 export interface StudiolineRail {
