+++ frontend/src/semantic/rx-tx-surface.ts
@@ -238,103 +238,3 @@ export const txDisabledReasons = (view: RadioViewModel): readonly DisabledReason
 let sequence = 0;
 /** Per-instance DOM id, so several mounted surfaces keep distinct aria targets. */
 export const nextSurfaceId = (): string => `rx-tx-${++sequence}`;
-
-/**
- * MOR-2031 — the design-language-agnostic half of TX state feedback.
- *
- * `fieldline` and `studioline` each render TX state as a rail plus a
- * key/slab control, and until this ticket each language's own
- * `state-feedback-renderer.ts` re-derived the identical fail-closed decision
- * over its own private copy of the logic — `resolveTxFeedbackState` is that
- * one decision, extracted. (The `KEY_TREATMENT` ordering below — the F3/N3
- * fix that makes `keyBlocked` lose to a louder session — was itself
- * hand-duplicated into both files in a single commit, MOR-1275.)
- *
- * `TxFeedbackState.rail` names the session BUCKET only, never a magnitude:
- * fieldline and studioline rank the six buckets differently (fieldline's
- * rail floods `pending` to a MID-tier 16px; studioline's rail thickens
- * `pending` to its TOP-tier 3px — both pinned in each language's own
- * `state-feedback-renderer.test.ts`), so a shared ordinal could not serve
- * both. Each language keeps its own `Record<TxFeedbackRail, …>` for
- * width/thickness/label/band, indexed by the bucket this function resolves.
- *
- * SAFETY INVARIANT R9, identical to `rfState`/`txSessionState` above: this
- * function DISPLAYS the App TX authority's conclusions; it never forms its
- * own. It takes a named struct of exactly four primitives — never a whole
- * `RendererViewModel` — so a fifth field cannot be smuggled into a TX
- * decision without changing this signature. Unrecognised values fail CLOSED
- * to the `'doubt'` rail, never to `'idle'`: "nothing is happening" is the
- * dangerous claim.
- */
-export type TxFeedbackRail = 'idle' | 'pending' | 'keyed' | 'releasing' | 'failed' | 'doubt';
-export type TxKeyTreatment = 'idle' | 'pending' | 'keyed' | 'fault' | 'blocked';
-/** Symbolic only — NOT a token value. Each language maps this to its own `tokens.rx.idle` / `tokens.tx.active` / `tokens.tx.tuning`. */
-export type TxFeedbackTone = 'rx-idle' | 'tx-active' | 'tx-tuning';
-
-export interface TxFeedbackState {
-  readonly rail: TxFeedbackRail;
-  readonly treatment: TxKeyTreatment;
-  /** Transmitter engaged: true for exactly the `keyed` and `releasing` rails. */
-  readonly keyed: boolean;
-  /** The resolved RX baseline (`rail === 'idle'`) — the only silent state. */
-  readonly quiet: boolean;
-  readonly tone: TxFeedbackTone;
-  readonly failed: boolean;
-  /** `` `TX FAULT: ${fault}` `` when `failed` and `fault` is non-empty, else `null`. Both languages build the same string; only WHERE they place it differs. */
-  readonly faultText: string | null;
-}
-
-/** True for exactly `TxSessionState`'s five members — reuses `SESSION_LABEL`'s keys rather than re-enumerating them a third time. */
-const isTxSessionState = (value: string): value is TxSessionState =>
-  Object.prototype.hasOwnProperty.call(SESSION_LABEL, value);
-
-/**
- * Session → key treatment, unconditional on `keyBlocked`. Keyed by the RAW
- * session, not the resolved rail: an unrecognised session must still fall
- * through to the `keyBlocked`/doubt branch in `resolveTxFeedbackState` below
- * rather than being caught here.
- */
-const KEY_TREATMENT: Record<string, TxKeyTreatment> = {
-  idle: 'idle', pending: 'pending', keyed: 'keyed', releasing: 'keyed', failed: 'fault',
-};
-
-/**
- * Reads a flat renderer field as a string, defaulting an absent or
- * non-string value to `''`. Shared by both design languages' state-feedback
- * renderers, which previously each defined this identically.
- */
-export const stringField = (
-  fields: Readonly<Record<string, string | number | boolean | null>>, key: string,
-): string => {
-  const value = fields[key];
-  return typeof value === 'string' ? value : '';
-};
-
-export function resolveTxFeedbackState(fields: {
-  rf: string; session: string; fault: string; keyBlocked: boolean;
-}): TxFeedbackState {
-  const { rf, session, fault, keyBlocked } = fields;
-  // An unrecognised session, or an idle session the authority cannot vouch
-  // for as genuinely receiving, both land on the doubt rail.
-  const rail: TxFeedbackRail = isTxSessionState(session) && (session !== 'idle' || rf === 'receiving')
-    ? session
-    : 'doubt';
-  const failed = session === 'failed';
-  const keyed = rail === 'keyed' || rail === 'releasing';
-  const quiet = rail === 'idle';
-  const tone: TxFeedbackTone = quiet ? 'rx-idle' : keyed || failed ? 'tx-active' : 'tx-tuning';
-
-  // Treatment follows the RESOLVED rail, not the raw session: a control that
-  // still looks idle under a doubt rail is the contradiction R9 exists to
-  // prevent. `keyBlocked` is checked only once the session itself has
-  // nothing louder to say — see `KEY_TREATMENT`'s own comment for why.
-  const held = KEY_TREATMENT[session];
-  const treatment: TxKeyTreatment = held !== undefined && held !== 'idle' ? held
-    : keyBlocked ? 'blocked'
-      : rail === 'doubt' ? 'pending' : 'idle';
-
-  return {
-    rail, treatment, keyed, quiet, tone, failed,
-    faultText: failed && fault ? `TX FAULT: ${fault}` : null,
-  };
-}
